backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic>=2.10,<3
python-dotenv==1.0.1

# Agents SDK (matches your notebook import: from agents import ...)
openai-agents==0.0.12